% Auto-generated: do not edit by hand
\name{''Filterbuilder}

\alias{''Filterbuilder}

\title{Filterbuilder component}

\description{

}

\usage{
''Filterbuilder(id=NULL, fields=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{fields}{Unnamed list. An key,value array of datafield with at least a datafield key, example [{dataField: 'Name',}, {dataField: 'Price',dataType: 'number',format: 'currency',}, {dataField: 'Current_Inventory',dataType: 'number',caption: 'Inventory',}]}

\item{value}{Unnamed list. builded array of combined queries}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

